Commodity refactoring.
getPrice => getPriceAmount (when we want the float)
getPrice => getPrice (When we want the object)
get => create (when they are creating objects not retrieving instance variables)

stack => VerticalViewContainer
row => HorizontalViewContainer

=== Simplified MarketTableContent. ===
1) We create rows now whenever we initialize the table
2) We do not re-create the table every time we ask for information
3) Rows are wrapped inside RowDataElement (your favorite builder pattern)

=== Decoupling markets ===
This is half way done. Tooltip part has not been touched at the moment and it is still tightly coupled together with MarketAPI. Ideally it should ask for information from MarketRepository. MarketRepository currently is not injected into classes but when I've figured out what to do with it, it probably will be.

1) Added wrapper for MarketAPI to simplify some of the actions on the object
2) Removed some the entanglement with Prices